added comments, and stopped glBufferSubData
The host CPU memory for the particle positions, the openCL memory for
those, and the openGL memory for those are all linked.
glBufferSubData is a call that I understood to tell openGL that the
openCL had updated the bytes.  Turns out this call is bad: it copies
out of date or random data into the openGL vectors.  Disabling this
fixed that problem.